gsdb refactor
